Separate doc pages, add file level docstrings
