+++ docs/code_documentation/data_reader.md
@@ -1,0 +1,1 @@
+::: sodym.data_reader
