+++ docs/code_documentation/mfa_definition.md
@@ -1,0 +1,1 @@
+::: sodym.mfa_definition
