+++ docs/code_documentation/mfa_system.md
@@ -1,0 +1,1 @@
+::: sodym.mfa_system
