--- docs/code_documentation/model_and_components.md
@@ -1,6 +1,0 @@
-::: sodym.mfa_system
-::: sodym.mfa_definition
-::: sodym.data_reader
-::: sodym.dimensions
-::: sodym.named_dim_arrays
-::: sodym.flow_helper
