+++ docs/code_documentation/named_dim_arrays.md
@@ -1,0 +1,3 @@
+::: sodym.named_dim_arrays
+::: sodym.named_dim_array_helper
+::: sodym.flow_helper
